chore: rename project to yaci
