+++ internal/output/json.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/liftedinit/cosmos-dump/internal/models"
+	"github.com/liftedinit/yaci/internal/models"
 )
 
 type JSONOutputHandler struct {
